Issue #40. Permutations refactored to Intuition library.
